# impala-bridge/openapi.yaml
openapi: 3.1.0
info:
  title: impala-bridge
  version: "0.0.0"
  summary: REST bridge between offline Payala payments and the Stellar network.
  description: |
    impala-bridge issues JWTs, persists dual-chain transactions, dispatches
    notifications via SMS/email/push/webhook, and streams events from Stellar
    Horizon and Payala.

    Authentication uses a two-token strategy:

    - **Refresh token** (14-day TTL) — obtained via username+password from `POST /token`.
      Used only to mint temporal tokens, never for API access.
    - **Temporal token** (1-hour TTL) — obtained from a refresh token via `POST /token`.
      Sent as `Authorization: Bearer <token>` on every authenticated request.

    All mutating endpoints enforce account ownership via `require_owner`. A
    `POST /logout` revokes the current token (the JTI is added to a Redis
    blacklist until it would have expired naturally).

    Rate limits apply to `/authenticate`, `/token`, `/mfa/verify`, `/transaction`,
    and `/card`: 10 requests per 60-second window, keyed by account id (and
    MFA type where applicable). Exceeding the limit returns HTTP 429 with a
    `Retry-After` header.

    Global request timeout is 30 s (HTTP 408 on exceed). Postgres has a
    per-connection `statement_timeout` of 60 s.
  contact:
    name: impala-bridge maintainers
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
...
  - url: https://{host}
...
security:
  - bearerAuth: []
paths:
...
  /authenticate:
    post:
      tags: [auth]
      summary: Register or verify a user with username+password.
      description: |
        If no credentials exist for the account, registers the user with the
        supplied password (Argon2 hashed). If credentials exist, verifies the
        password. Returns a generic "Invalid credentials" for both missing
        accounts and wrong passwords (account-enumeration defense). Rate
        limited per account id (10/60 s) and per client source (30/60 s,
        shared across all pre-auth endpoints); lockout after 5 wrong
        passwords per (account, client source) for 15 minutes — unknown
        accounts never count toward it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthenticateRequest' }
      responses:
        '200':
          description: Authentication outcome (success may be false with 200).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticateResponse' }
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /managed-account/generate:
    post:
      tags: [managed-account]
      summary: Generate a custodial Stellar account (owner-scoped).
      description: |
        Creates a fresh Stellar keypair server-side, protects the seed with the
        configured backend, and stores it. Only the public `G...` address is
        returned — the seed never leaves the bridge. Rate limited per account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateManagedAccountRequest' }
      responses:
        '200':
          description: Managed account created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManagedAccountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /managed-account/sign:
    post:
      tags: [managed-account]
      summary: Sign and submit a native payment from a custodial account (owner-scoped).
      description: |
        Loads the protected seed for the owner's account, signs a native (XLM)
        payment to `destination`, submits it to the network, and records the
        resulting transaction. Synchronous and server-only so a retry cannot
        double-submit. Rate limited per account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignSubmitRequest' }
      responses:
        '200':
          description: Payment signed and submitted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignSubmitResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Temporal token (1-hour TTL). Obtain via `POST /token` with a valid
        refresh token. Send as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: impala_session
      description: |
        Browser session cookie (HttpOnly), established by `POST /session/login`
        or `POST /auth/okta` with `cookie_mode: true`. In production the cookie
        is `__Host-`-prefixed (`__Host-impala_session`). Unsafe-method requests
        additionally require the CSRF synchronizer token in the `X-CSRF-Token`
        header.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Token is valid but the resource belongs to another account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Request conflicts with the current state (e.g. last admin).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "unauthorized" }
            message: { type: string }
          required: [code, message]
      required: [error]
    AuthenticateRequest:
      type: object
      properties:
        account_id: { type: string }
        password: { type: string, minLength: 8 }
      required: [account_id, password]
    AuthenticateResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        action:
          type: string
          enum: ["", registered, authenticated]
...
    GenerateManagedAccountRequest:
      type: object
      properties:
        payala_account_id: { type: string }
        first_name: { type: string, maxLength: 64 }
        middle_name: { type: string, nullable: true }
        last_name: { type: string, maxLength: 64 }
        nickname: { type: string, nullable: true }
        affiliation: { type: string, nullable: true }
        gender: { type: string, nullable: true }
      required: [payala_account_id, first_name, last_name]
...
    ManagedAccountResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        stellar_account_id: { type: string, nullable: true }
    SignSubmitRequest:
      type: object
      properties:
        payala_account_id: { type: string }
        destination: { type: string, description: Destination Stellar G-address. }
        amount: { type: string, description: Native (XLM) amount as a decimal string. }
        memo: { type: string, nullable: true }
        fee: { type: integer, nullable: true, description: Base fee in stroops. }
      required: [payala_account_id, destination, amount]
    SignSubmitResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        stellar_hash: { type: string, nullable: true }
        btxid: { type: string, format: uuid, nullable: true }
...